api: move music curation types into owning pair files
